WIP refactoring server side code
